Split resolver collection enum type tests
